Add slice preprocess for dota and yolo.

+++ ultralytics/nn/modules/block.py
@@ -2025,7 +2025,9 @@ class SAVPE(nn.Module):
         vp = vp.reshape(B, Q, 1, -1)
 
         score = y * vp + torch.logical_not(vp) * torch.finfo(y.dtype).min
-        score = F.softmax(score, dim=-1).to(y.dtype)
+
+        score = F.softmax(score, dim=-1, dtype=torch.float).to(score.dtype)
+
         aggregated = score.transpose(-2, -3) @ x.reshape(B, self.c, C // self.c, -1).transpose(-1, -2)
 
         return F.normalize(aggregated.transpose(-2, -3).reshape(B, Q, -1), dim=-1, p=2)
